+++ test/postgresql_test_helper.rb
@@ -2,18 +2,7 @@
 
 require 'test_helper'
 
-PostgresqlDB = Sequel.connect('postgres:///test')
-
-Sequel::Model.db = PostgresqlDB
-
-class Thing < Sequel::Model
-end
-
-class InheritedThing < Thing
-end
-
-class Master < Sequel::Model
-end
+PostgresqlDB = Sequel.connect('postgres://postgres:postgres@127.0.0.1:5432/test')
 
 module PostgresqlTestHelper
   def recreate_table
@@ -24,5 +13,13 @@ module PostgresqlTestHelper
     PostgresqlDB.run 'DROP SEQUENCE IF EXISTS c'
     sql = 'CREATE TABLE things (id SERIAL PRIMARY KEY, slug VARCHAR(255), quantity INTEGER DEFAULT 0);'
     PostgresqlDB.run sql
+  end
+
+  def with_migration(&block)
+    migration_class = Sequel::Migration
+
+    Sequel::Model.db = PostgresqlDB
+
+    Class.new(migration_class, &block).new(PostgresqlDB)
   end
 end